derive_more_impl::from

Struct Expansion

struct Expansion<'a> {
    attrs: Option<&'a FieldConversion>,
    ident: &'a Ident,
    variant: Option<&'a Ident>,
    fields: &'a Fields,
    generics: &'a Generics,
    has_explicit_from: bool,
}

Expansion of a macro for generating From implementation of a struct or enum.

Fields
attrs: Option<&'a FieldConversion>

From attributes.

As a VariantAttribute is superset of a StructAttribute, we use it for both derives.

ident: &'a Ident

Struct or enum syn::Ident.

variant: Option<&'a Ident>

Variant syn::Ident in case of enum expansion.

fields: &'a Fields

Struct or variant syn::Fields.

generics: &'a Generics

Struct or enum syn::Generics.

has_explicit_from: bool

Indicator whether one of the enum variants has VariantAttribute::Empty, VariantAttribute::Types or VariantAttribute::Forward.

Always false for structs.

Implementations

impl Expansion<'_>

fn expand(&self) -> Result<TokenStream>

Expands From implementations for a struct or an enum variant.

fn expand_fields( &self, wrap: impl FnMut(Option<&Ident>, &Type, Option<Index>) -> TokenStream, ) -> TokenStream

Expands fields initialization wrapped into token::Braces in case of syn::FieldsNamed, or token::Paren in case of syn::FieldsUnnamed.