derive_more_impl::ops

Struct AssignStructuralExpansion

struct AssignStructuralExpansion<'i> {
    trait_ty: Ident,
    method_ident: Ident,
    self_ty: (&'i Ident, &'i Generics),
    variants: Vec<(Option<&'i Ident>, &'i Fields, HashSet<usize, DeterministicState>)>,
    is_enum: bool,
}

Expansion of a macro for generating a structural trait implementation with a &mut self method receiver for an enum or a struct.

Fields
trait_ty: Ident

syn::Ident of the implemented trait.

method_ident: Ident

syn::Ident and syn::Receiver of the implemented method in trait.

self_ty: (&'i Ident, &'i Generics)

syn::Ident and syn::Generics of the implementor enum/struct.

variants: Vec<(Option<&'i Ident>, &'i Fields, HashSet<usize, DeterministicState>)>

syn::Fields of the enum/struct to be used in this AssignStructuralExpansion.

is_enum: bool

Indicator whether this expansion is for an enum.

Implementations

impl AssignStructuralExpansion<'_>

fn body(&self) -> TokenStream

Generates body of the method implementation for this StructuralExpansion.