derive_more_impl

Module utils

Modules

attr 🔒

either 🔒

fields_ext 🔒

generics_search 🔒

pattern_matching 🔒

Helper extensions of syn types for pattern matching code generation.

polyfill 🔒

Polyfill for syn 1.x AST.

spanning 🔒

structural_inclusion 🔒

Helper extensions of syn types for checking structural inclusion.

Structs

AttrParams

DeterministicState

FullMetaInfo

MetaInfo

MultiFieldData

MultiVariantData

SingleFieldData

State

Enums

DeriveType

RefType

Functions

add_extra_generic_param

add_extra_generic_type_param

add_extra_ty_param_bound

add_extra_type_param_bound_op_output

add_extra_where_clauses

add_where_clauses_for_new_ident

field_idents

get_field_types

get_field_types_iter

get_if_type_parameter_used_in_type

get_meta_info 🔒

is_type_parameter_used_in_type

named_to_vec

numbered_vars

panic_one_field 🔒

parse_punctuated_nested_meta 🔒

unnamed_to_vec

Type Aliases

HashMap

HashSet