derive_more_impl::utils::attr

Module field_conversion

Enums

FieldConversion 🔒

Representation of an attribute, specifying which conversions should be generated: either forwarded via a blanket impl, or direct for concrete specified types.

Type Aliases

Untyped 🔒

Untyped analogue of a FieldConversion, recreating its type structure via Either.