Crate derive_more_impl

derive_more-impl

This crate is an implementation detail of the derive_more. If you found this crate by accident you’re probably looking for one of the following pages of derive_more:

crates.io

docs.rs

GitHub

Modules

deref 🔒

deref_mut 🔒

from 🔒

Implementation of a From derive macro.

ops 🔒

Implementations of ops-related derive macros.

utils 🔒

Macros

create_derive 🔒

Traits

Output 🔒

Derive Macros

Add

What #[derive(Add)] generates

AddAssign

What #[derive(AddAssign)] generates

BitAnd

What #[derive(Add)] generates

BitAndAssign

What #[derive(AddAssign)] generates

BitOr

What #[derive(Add)] generates

BitOrAssign

What #[derive(AddAssign)] generates

BitXor

What #[derive(Add)] generates

BitXorAssign

What #[derive(AddAssign)] generates

Deref

Using #[derive(Deref)]

DerefMut

What #[derive(DerefMut)] generates

From

What #[derive(From)] generates

Sub

What #[derive(Add)] generates

SubAssign

What #[derive(AddAssign)] generates