derive_more_impl

Module utils

Modules

attr ğŸ”’

either ğŸ”’

fields_ext ğŸ”’

generics_search ğŸ”’

pattern_matching ğŸ”’

Helper extensions of syn types for pattern matching code generation.

polyfill ğŸ”’

Polyfill for syn 1.x AST.

spanning ğŸ”’

structural_inclusion ğŸ”’

Helper extensions of syn types for checking structural inclusion.

Structs

AttrParams

DeterministicState

FullMetaInfo

MetaInfo

MultiFieldData

MultiVariantData

SingleFieldData

State

Enums

DeriveType

RefType

Functions

add_extra_generic_param

add_extra_generic_type_param

add_extra_ty_param_bound

add_extra_type_param_bound_op_output

add_extra_where_clauses

add_where_clauses_for_new_ident

field_idents

get_field_types

get_field_types_iter

get_if_type_parameter_used_in_type

get_meta_info ğŸ”’

is_type_parameter_used_in_type

named_to_vec

numbered_vars

panic_one_field ğŸ”’

parse_punctuated_nested_meta ğŸ”’

unnamed_to_vec

Type Aliases

HashMap

HashSet